Read Description
 I fixed the connection string. Our program can now access our database, this was very time consuming but I finally got it to work.
To achieve that, I implemented a "scaffold" procedure as well as installed some needed "sqlserver" packages.
Issue-fixed:
(Initially, there was an issue when I tried to run the program, it kept getting stuck at @await Html.PartialAsync("_LoginPartial") in Views/Shared/_Layout.cshtml. When I commented it out, the program run successfully, but it didn't have the Login option. I was afraid this was a route we would have to take but I took a break from the project, came back, ran it, with the @await Html.PartialAsync call & it worked! Not sure what happened, quite mysterious. I was stuck on that for more than an hour and it suddenly worked)
By the scaffold proceudre, ".cs" files were automatically created for our tables through this procedure. The tables without primary keys were not successfully added, I have a screenshot  summary of those tables. If needed, we can manually add ".cs" files  for any of the reference tables, now that we have templates to refer back to from the other ".cs" files.