Edit Catalog, Approriate Roles Assigned
Customer and Employees who are signed in are able to checkout and view info of books.
Additionally Employees may edit the Catalog *******
Non-members / people who are not signed in can see:
The catalog. When they try to checkout, they are redirected to the login page where they can either login or register.